[3022] Make improvements according to PR comments
