+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/suggestion/list/SuggestionListUi.kt
@@ -4,7 +4,18 @@ import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.ui.suggestion.Suggestions
 
 @ExperimentalStreamChatApi
+/**
+ * Interface used for communication with the suggestion list view.
+ */
 public interface SuggestionListUi {
+
+    /**
+     * Renders the suggestion list View.
+     */
     public fun renderSuggestions(suggestions: Suggestions)
+
+    /**
+     * Shows if the suggestion list is currently visible.
+     */
     public fun isSuggestionListVisible(): Boolean
 }
